fix phpstan level 5

+++ src/Builder/adebipe/Router/RouteKeeperBuilder.php
@@ -58,7 +58,7 @@ class RouteKeeperBuilder implements BuilderInterface
             str_shuffle(
                 str_repeat(
                     $x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ',
-                    ceil($length / strlen($x))
+                    (int)ceil($length / strlen($x))
                 )
             ),
             1,
